requirements.txt
openai>=1.0.0
anthropic>=0.5.0
mistralai>=1.0.0
groq>=0.3.0
google-generativeai>=0.3.0
cohere>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
plotly>=5.15.0
numpy>=1.24.0
pandas>=2.0.0
